requirements.txt
python-telegram-bot
google-genai
python-dotenv
requests
